Support C code generate for GIGO hw
The concept here is to generate related c code wrapper which will call
gigo_xxx function (e.g. gigo_conv2d).

In the gigo_xxx function, it is expected to transfer buffer from host to
GIGO hw, and perform operations on it. After that, the result will be
copied back to host.

In this commit, only `conv2d` is wrapped.